backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
git+https://github.com/openai/whisper.git
torch==2.0.1
torchaudio==2.0.2
python-multipart==0.0.6
ffmpeg-python==0.2.0
python-dotenv==1.0.0
pydantic==1.10.7
